refactor: profile components
